requirements.txt
# FastAPI framework
fastapi==0.111.0

# ASGI server for FastAPI
uvicorn[standard]==0.29.0

# OCR processing
pytesseract==0.3.10
pillow==10.3.0

# For fetching image from URL
requests==2.31.0
